requirements.txt
Django==1.9
django-angular==0.7.15
django-bower==5.0.4
django-filter==0.11.0
djangorestframework==3.3.2
requests==2.9.1
six==1.10.0
virtualenv==13.1.2
wheel==0.24.0
